% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_pathways.R
\name{assess_pathways}
\alias{assess_pathways}
\title{Quality assessment of input movement data.}
\usage{
assess_pathways(pathways)
}
\arguments{
\item{pathways}{The named list "pathways" in the output of function read_pathways.}
}
\value{
A logical value indicating whether the data pass (TRUE) or fail (FALSE)
this assessment.
}
\description{
This function evaluates the quality of movement data under four
criteria: 1. Time_start <= Time_end at each location; 2. location uniqueness;
3. pathway integrity; 4. pathway uniqueness. See Subsection "Requirements for
 the quality of input movement data" on github.com/wanyuac/pathopath for details.
}
\author{
Yu Wan, \email{yu.wan@liverpool.ac.uk}
}
