% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pathways.R
\name{build_pathways}
\alias{build_pathways}
\title{Build patient pathways from movements}
\usage{
build_pathways(movements)
}
\arguments{
\item{movements}{Output tibble of function import_movements}
}
\value{
A named list with each element named by a unique subject ID and the element's
value is a tibble of this subject's movements, sorted by Time_in from the oldest to
the latest.
}
\description{
This function takes as input the output tibble of import_movements. The tibble
has four columns: Subject, Location, Time_in, and Time_out.
}
