% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_by_genetic_dists.R
\name{clustering_by_genetic_dists}
\alias{clustering_by_genetic_dists}
\title{Clustering subjects by pairwise genetic distances}
\usage{
clustering_by_genetic_dists(
  matrix_tsv = NULL,
  threshold = 15,
  excl_subjects = ""
)
}
\arguments{
\item{matrix_tsv}{Path to the input distance matrix in the TSV format.}

\item{threshold}{Threshold of SNP distances. Integer. Default: 15.}

\item{excl_subjects}{An optional vector of subject names for exclusion from the clustering analysis.}
}
\value{
A tibble comprising three columns: Subject, Threshold, and Genetic_cluster
}
\description{
This function takes as input a matrix of genetic distances, which can be a SNP-distance matrix (Hamming
distances) generated from a sequence alignment by software snp-dists, and clusters subjects based on a distance threshold.
}
\author{
Yu Wan, \email{yu.wan@liverpool.ac.uk}
}
