% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dn_clustering.R
\name{dn_clustering}
\alias{dn_clustering}
\title{Clustering of subjects in a distance network as graph components}
\usage{
dn_clustering(
  mat = NULL,
  thresholds = c(5L, 10L, 15L, 20L, 50L),
  excl_subjects = ""
)
}
\arguments{
\item{mat}{Path to the input distance matrix in the TSV format.}

\item{excl_subjects}{An optional vector of subject names for exclusion from the clustering analysis.}

\item{threshold}{Positive integers for distance thresholds. Default: c(5, 10, 15, 20, 50).}
}
\value{
A Clusters object comprising the following slots: (1) data frame "clusters" consisting of columns "Subject",
"Cluster_0", ..., "Cluster_n", where n represents the largest distance threshold in the vector "thresholds"; (2) data
frame "cluster_counts" of two columns "Threshold", "Cluster_count"; (3) an empty list "tree", intentionally left empty
for consistency with h_clustering output; and (4) symmetric matrix "distances", storing the input distance matrix; if
the excl_subjects parameter is specified, certain subjects are excluded from this matrix.
}
\description{
This function takes as input a matrix of Hamming distances, which can be a SNP-distance matrix generated
from a sequence alignment by software snp-dists, and a vector of distance thresholds. The distance matrix is equivalent
to an all-to-all bidirectional distance network. Under each distance threshold, this function prunes edges that exceeds
the threshold in the network and identifies components in the remaining network as clusters of nodes (subjects). Comparing
to the approach of complete-linkage hierarchical clustering implemented in function h_clustering, this approach walks
through a contact network to identify maximal clusters of nodes based on a given maximum stepwise distance. The prefix
"dn" of the function's name stands for "distance-network".

Users do not need to include zero in the threshold vector, since this function always determines clusters of genetically
identical subjects, generating column "Cluster_0" in the output tibble of cluster identifiers.
}
\author{
Mohammad Saiful Islam Sajib, \email{saiful.sajib@chrfbd.org}

Yu Wan, \email{yu.wan@liverpool.ac.uk}
}
