% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_clustering.R
\name{hd_clustering}
\alias{hd_clustering}
\title{Complete-linkage hierachical clustering of subjects by Hamming distances}
\usage{
hd_clustering(
  mat = NULL,
  thresholds = c(5L, 10L, 15L, 20L, 50L),
  excl_subjects = ""
)
}
\arguments{
\item{mat}{Path to the input distance matrix in the TSV format.}

\item{excl_subjects}{An optional vector of subject names for exclusion from the clustering analysis.}

\item{threshold}{Threshold of SNP distances. Integer. Default: c(5, 10, 15, 20, 50).}
}
\value{
A Clusters object comprising the following slots: (1) data frame "clusters" consisting of columns "Subject",
"Cluster_0", ..., "Cluster_n", where n represents the largest distance threshold in the vector "thresholds"; (2) data
frame "cluster_counts" of two columns "Threshold", "Cluster_count"; (3) named list "tree", which consists of a dendrogram
"tree" (rooted by the clustering algorithm) and a character field "method" (complete linkage); and (4) symmetric matrix
"distances", storing the input distance matrix.
}
\description{
This function takes as input a matrix of Hamming distances, which can be a SNP-distance matrix generated
from a sequence alignment by software snp-dists, and a vector of distance thresholds (positive integers). It clusters
subjects using complete-linkage hierarchical clustering and then partitions subjects into clusters under each distance
threshold. Users do not need to include zero in the threshold vector, since this function always determines clusters
of genetically identical subjects, generating column "Cluster_0" in the output tibble of cluster identifiers.
}
\author{
Yu Wan, \email{yu.wan@liverpool.ac.uk}
}
