% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_movements.R
\name{import_movements}
\alias{import_movements}
\title{Import the movements spreadsheet}
\usage{
import_movements(movement_tsv)
}
\arguments{
\item{movement_tsv}{Path to a spreadsheet of subjects' location records in the tab-delimited format (TSV).}
}
\value{
Tibble movements
}
\description{
Read the TSV file recording subjects' locations and the period at each location.
Required columns: Subject, Location, Time_in, Time_out, Admission.
}
\author{
Yu Wan, \email{yu.wan@liverpool.ac.uk}

Mohammad Saiful Islam Sajib, \email{saiful.sajib@chrfbd.org}
}
