% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_clustering.R
\name{sample_clustering}
\alias{sample_clustering}
\title{Clustering of microbiological samples by Hamming distances}
\usage{
sample_clustering(
  mat = NULL,
  method = "hc",
  thresholds = c(0L, 5L, 10L, 15L, 20L, 50L),
  excl_samples = ""
)
}
\arguments{
\item{mat}{Path to the input distance matrix in the TSV format.}

\item{method}{Clustering method: "hc", complete-linkage hierarchical clustering (default); "dn", distance-network clustering.}

\item{thresholds}{Positive integers for distance thresholds. Default: c(0L, 5L, 10L, 15L, 20L, 50L).}

\item{excl_samples}{An optional vector of sample names to be excluded from the clustering analysis.}
}
\value{
A Clusters object comprising the following slots: (1) data frame "clusters" consisting of columns "Sample",
"Cluster_0", ..., "Cluster_n", where n represents the largest distance threshold in the vector "thresholds"; (2) data
frame "cluster_counts" of two columns "Threshold", "Cluster_count"; (3) named list "tree", which consists of a dendrogram
"tree" (rooted by the clustering algorithm) and a character field "method" (complete linkage) when method = "hc", otherwise,
an empty list object; and (4) symmetric matrix "distances", storing the input distance matrix; if the excl_samples parameter is
specified, certain samples are excluded from this matrix.
}
\description{
This function takes as input a matrix of Hamming distances, which can be a SNP-distance matrix generated
from a sequence alignment by software snp-dists for pathogens, and a vector of distance thresholds. It clusters samples
using either complete-linkage hierarchical clustering or a distance-network-based approach. In the latter, The distance
matrix is treated as an equivalent all-to-all bidirectional distance network. Under each distance threshold, this function
prunes edges that exceeds the threshold in the network and identifies components in the remaining network as clusters of
nodes representing samples. Comparing to the complete-linkage hierarchical clustering, this approach walks through a
contact network to identify maximal clusters of nodes based on a given maximum stepwise distance.

After clustering, the function partitions samples into clusters under each distance threshold. This function always determines
clusters of genetically identical samples (namely, with the threshold of zero), generating column "Cluster_0" in the output
data frame of cluster identifiers.
}
\author{
Yu Wan, \email{yu.wan@liverpool.ac.uk}
}
