% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_movements.R
\name{add_movements}
\alias{add_movements}
\title{Incorporation of additional patient movements into existing results}
\usage{
add_movements(
  movement_table = NULL,
  genotype_table = NULL,
  previous_results = NULL
)
}
\arguments{
\item{movement_table}{Path to a tab-delimited spreadsheet of new movements to be incorporated into the existing results.}

\item{genotype_table}{Path to a tab-delimited spreadsheet of pathogen genotypes. Given the complexity of genotypical
data, currently users need to provide curated genotypes for all samples, including previously analysed samples.}

\item{previous_results}{A previous output object of pathopath.}
}
\description{
This function adds patient movements into existing results without requiring
reprocessing the whole dataset.
}
\author{
Yu Wan, \email{yu.wan@liverpool.ac.uk}
}
