% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_contacts.R
\name{compute_contacts}
\alias{compute_contacts}
\title{Determine direct and indirect contacts from pathways}
\usage{
compute_contacts(pathways, dt = 3)
}
\arguments{
\item{pathways}{Named list "pathways" in the output of function read_pathways.}

\item{dt}{Delta t, ± dt days to determine an indirect contact. Set it to zero to turn off the detection of indirect contacts.
Default: 3.}
}
\value{
Edges tibble with from, to, contact_type, direct_days, indirect_days, duration, delta_t, d0. Attribute 'resolution' is set.
}
\description{
A direct contact is defined as an overlap between pathways of distinct subjects at the same location and on the same day.
An indirect contact is defined as an overlap between pathways of distinct subjects at the same location and not overlap
in time unless allowing ±dt days.
}
