% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_contacts.R
\name{compute_contacts}
\alias{compute_contacts}
\title{Compute direct and indirect contacts from pathways}
\usage{
compute_contacts(pathways, indirect_cutoff = NULL, dt = NULL, d0 = 1)
}
\arguments{
\item{pathways}{Output of function build_pathways}

\item{indirect_cutoff}{Numeric; alias is `dt`. If neither is set, defaults to 0.}

\item{dt}{Delta t, ± dt days to determine an indirect contact. If both are set, `indirect_cutoff` wins.}

\item{d0}{Duration for a point overlap in time. For example, if two subject were only at the same location
for one day or even one hour. Default: 1 (day).}
}
\value{
Edges tibble with from, to, contact_type, direct_days, indirect_days, duration, delta_t, d0. Attribute 'resolution' is set.
}
\description{
A direct contact is defined as an overlap between pathways of distinct subjects at the same location and on the same day.
An indirect contact is defined as an overlap between pathways of distinct subjects at the same location and not overlap
in time unless allowing ±dt days.
}
