% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{create_network}
\alias{create_network}
\title{Create undirected, Cytoscape-compatible network files from a contact summary}
\usage{
create_network(contact_summary, samples = NULL)
}
\arguments{
\item{contact_summary}{The tibble or data frame of contact summaries in pathopath's output (slot "summary").}

\item{samples}{Optional path to a tab-delimited spreadsheet of microbiological/pathological data, with three mandatory columns
Sample, Subject, and Pathway, followed by optional variable columns such as sample metadata, genotypes, or phenotypes. The
sample data will be added to node attributes for network visualisation and analysis.}
}
\value{
A network object with two tibble slots: V for nodes and E for edges.
}
\description{
This function creates a node table and edge table that can be visualised in Cytoscape or other compatible
software for network analysis at the level of movement pathways.
}
\author{
Yu Wan, \email{yu.wan@liverpool.ac.uk}

Mohammad Saiful Islam Sajib, \email{saiful.sajib@chrfbd.org}
}
