% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathopath.R
\name{pathopath}
\alias{pathopath}
\title{Construct a contact network from epidemiological data}
\usage{
pathopath(pathway_data = NULL)
}
\arguments{
\item{pathway_data}{Path to a spreadsheet of subjects' movement records arranged in pathways in the tab-delimited
format. The Location column stores location accessions, which are unique identifiers of locations at a user-specified level.}
}
\value{
A named list of two elements "pathways" and "contacts".
}
\description{
This is the main function of the Pathopath package
}
\author{
Yu Wan, \email{yu.wan@liverpool.ac.uk}

Mohammad Saiful Islam Sajib, \email{saiful.sajib@chrfbd.org}
}
