% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathopath.R
\name{pathopath}
\alias{pathopath}
\title{Construct a contact network from epidemiological data}
\usage{
pathopath(movement_data = NULL)
}
\arguments{
\item{movement_data}{Path to a spreadsheet of subjects' movement records in the tab-delimited format. The Location
column stores location accessions, which are unique identifiers of locations at a user-specified level.}
}
\value{
An object of Class ContactNetwork
}
\description{
This is the main function of the Pathopath package
}
\author{
Yu Wan, \email{yu.wan@liverpool.ac.uk}

Mohammad Saiful Islam Sajib, \email{saiful.sajib@chrfbd.org}
}
