% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathopath.R
\name{pathopath}
\alias{pathopath}
\title{Construct a contact network from epidemiological data}
\usage{
pathopath(pathway_data = NULL, dt = 3)
}
\arguments{
\item{pathway_data}{Path to a spreadsheet of subjects' movement records arranged in pathways in the tab-delimited
format. The Location column stores location accessions, which are unique identifiers of locations at a user-specified level.}

\item{dt}{Delta t, ±dt days (inclusive) to determine an indirect contact. Set it to zero to turn off the detection of
indirect contacts. Default: 3.}
}
\value{
A Pathopath object of three slots: (1) pathways, a list of pathway tibbles named by subject names; (2) migrations,
a tibble summarising the number of migrations per pathway; (3) contacts, a tibble of contact status (Direct, Indirect, and None)
between pathways of different subjects across shared locations.
}
\description{
This is the main function of the Pathopath package
}
\author{
Yu Wan, \email{yu.wan@liverpool.ac.uk}

Mohammad Saiful Islam Sajib, \email{saiful.sajib@chrfbd.org}
}
