% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathopath.R
\name{pathopath}
\alias{pathopath}
\title{Construct a contact network from epidemiological data}
\usage{
pathopath(epi_tsv = NULL)
}
\arguments{
\item{epi_tsv}{Path to a spreadsheet of epidemiological data in the tab-delimited format (TSV)}
}
\value{
An object of Class ContactNetwork
}
\description{
This is the main function of the Pathopath package
}
\author{
Yu Wan, \email{yu.wan@liverpool.ac.uk}

Mohammad Saiful Islam Sajib, \email{saiful.sajib@chrfbd.org}
}
