% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathopath.R
\name{pathopath}
\alias{pathopath}
\title{Construct a contact network from epidemiological data}
\usage{
pathopath(movements = NULL, samples = NULL, dt = 3)
}
\arguments{
\item{movements}{Path to a tab-delimited spreadsheet of subjects' movement records arranged in pathways. The Location
column stores location accessions, which are unique identifiers of locations at a user-specified level. The pathway accessions
must be unique across the input data. Note that by definition, subject and pathway accessions are in
one-to-one mapping.}

\item{samples}{Optional path to a tab-delimited spreadsheet of microbiological/pathological data, with three mandatory columns
Sample, Subject, Pathway, followed by optional variable columns such as sample metadata, genotypes, or phenotypes.The sample
data will be added to node attributes for network visualisation and analysis.}

\item{dt}{Delta t, ±dt days (inclusive) to determine an indirect contact. Set it to zero to turn off the detection of
indirect contacts. Default: 3.}
}
\value{
A Pathopath object of six slots: (1) pathways, a list of pathway tibbles named by subject names; (2) migrations,
a tibble summarising the number of migrations per pathway; (3) contacts, a tibble of contact status (Direct, Indirect, and None)
between pathways of different subjects across shared locations; (4) summary, a tibble summarising pathway-level direct and indirect
contacts for each unique combination of Subject_1, Subject_2, Pathway_1, and Pathway_2; (5) network, a Network object with two slots
- V for nodes and E for edges - for exportation as node and edge tables compatible with Cytoscape; (6) parameters, a named list
storing arguments of the pathopath function for reproducibility and recalculation for contacts.
}
\description{
This is the main function of the Pathopath package
}
\author{
Yu Wan, \email{yu.wan@liverpool.ac.uk}

Mohammad Saiful Islam Sajib, \email{saiful.sajib@chrfbd.org}
}
