% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pathways.R
\name{read_pathways}
\alias{read_pathways}
\title{Read the pathways spreadsheet}
\usage{
read_pathways(pathway_data, count_migrations = TRUE)
}
\arguments{
\item{pathway_data}{Path to a spreadsheet of user-defined pathways in the tab-delimited format (TSV).}

\item{count_migrations}{A logical switch turn on/off counting the number of migrations per pathway. Default: TRUE.}
}
\value{
An object of Class pathways, which consists of a list "pathways" with subject names as indices and a data
frame "migrations". The migration tibble is NULL when argument count_migrations = FALSE.
}
\description{
Read the TSV file recording subjects' pathway identifiers, locations, and the period at each location. Pathways are
defined by users. For instance, a pathway can be an admission of a patient. Required columns: (1) Subject: unique
subject identifier; (2) Pathway: unique pathway identifier; (3) Location: unique location identifier; (4) Time_start:
when the subject appeared at a specific location; (5) Time_end: when the subject left this location.
}
\author{
Yu Wan, \email{yu.wan@liverpool.ac.uk}

Mohammad Saiful Islam Sajib, \email{saiful.sajib@chrfbd.org}
}
